refactor: move states and screens into view
