+++ core/src/main/kotlin/com/tdt4240/group3/controller/PauseController.kt
@@ -3,10 +3,10 @@ package com.tdt4240.group3.controller
 import com.tdt4240.group3.game.playstate.PlaySubState
 import com.tdt4240.group3.controller.systems.TurnSystem
 import com.tdt4240.group3.model.components.TeamComponent
-import com.tdt4240.group3.screens.PlayScreen
-import com.tdt4240.group3.states.playstate.PauseState
-import com.tdt4240.group3.states.playstate.PlayerTurnState
-import com.tdt4240.group3.states.playstate.EnemyTurnState
+import com.tdt4240.group3.view.screens.PlayScreen
+import com.tdt4240.group3.view.states.PauseState
+import com.tdt4240.group3.view.states.PlayerTurnState
+import com.tdt4240.group3.view.states.EnemyTurnState
 
 class PauseController(
     private val turnSystem: TurnSystem,
