+++ core/src/main/kotlin/com/tdt4240/group3/controller/PlayController.kt
@@ -1,0 +1,6 @@
+package com.tdt4240.group3.controller
+
+class PlayController {
+    // initialize screen og config
+    // flytt kode fra screens hit
+}
