+++ core/src/main/kotlin/com/tdt4240/group3/controller/TurnController.kt
@@ -2,9 +2,9 @@ package com.tdt4240.group3.controller
 
 import com.tdt4240.group3.model.components.TeamComponent
 import com.tdt4240.group3.controller.systems.TurnSystem
-import com.tdt4240.group3.screens.PlayScreen
-import com.tdt4240.group3.states.playstate.EnemyTurnState
-import com.tdt4240.group3.states.playstate.PlayerTurnState
+import com.tdt4240.group3.view.screens.PlayScreen
+import com.tdt4240.group3.view.states.EnemyTurnState
+import com.tdt4240.group3.view.states.PlayerTurnState
 
 class TurnController(
     private val turnSystem: TurnSystem,
