+++ core/src/main/kotlin/com/tdt4240/group3/view/screens/HowToPlayScreen.kt
@@ -1,4 +1,4 @@
-package com.tdt4240.group3.screens
+package com.tdt4240.group3.view.screens
 
 import com.badlogic.gdx.Gdx
 import com.badlogic.gdx.Input
