+++ core/src/main/kotlin/com/tdt4240/group3/view/states/EnemyTurnState.kt
@@ -1,10 +1,9 @@
-package com.tdt4240.group3.states.playstate
+package com.tdt4240.group3.view.states
 
 import com.badlogic.gdx.Gdx
 import com.badlogic.gdx.Input
 import com.tdt4240.group3.game.playstate.PlaySubState
-import com.tdt4240.group3.screens.PlayScreen
-import ktx.app.clearScreen
+import com.tdt4240.group3.view.screens.PlayScreen
 
 class EnemyTurnState : PlaySubState {
     override val backgroundColor = Triple(0.35f, 0.1f, 0.1f)  // red
