+++ core/src/main/kotlin/com/tdt4240/group3/view/states/PlayerTurnState.kt
@@ -1,10 +1,9 @@
-package com.tdt4240.group3.states.playstate
+package com.tdt4240.group3.view.states
 
 import com.tdt4240.group3.game.playstate.PlaySubState
-import com.tdt4240.group3.screens.PlayScreen
+import com.tdt4240.group3.view.screens.PlayScreen
 import com.badlogic.gdx.Gdx
 import com.badlogic.gdx.Input
-import ktx.app.clearScreen
 
 class PlayerTurnState : PlaySubState {
     override val backgroundColor = Triple(0.1f, 0.35f, 0.1f)  // green
